==> machine_learning/machine_learning_offline/ds_competition/1.data_cleaning.ipynb <==
# 数据分析与数据清洗

因为按照常理助学金是给家庭有经济困难的学生的，而学生的消费情况比较能反映学生的经济状况，所以本次探索会先清洗出学生的各类消费情况及总消费。
学习成绩也应该会对奖学金的评定有一定的影响，以我在本科读书时的经验来看，成绩好的能拿奖学金一般不会申请或者不能通过助学金的审批，而曾经太差也会影响，所以我第二步准备将学生成绩清洗为在学院的排名，及学院人数这两个变量。

### 步骤
#### step1 清晰学生饭卡数据

#### step2 清洗学生学习成绩数据

## 处理学生消费数据

import pandas as pd 
import warnings 
warnings.filterwarnings("ignore")
names_card = [ '学生id','消费类别','消费地点','消费方式','消费时间','消费金额','剩余金额']
# 导入学生饭卡数据
card_train = pd.read_csv("../input/card_train.txt",header=None,encoding='utf-8',names = names_card,index_col = False)
card_test = pd.read_csv("../input/card_test.txt",header=None,encoding='utf-8',names = names_card,index_col = False)
card_data = pd.concat([card_train,card_test])

names_score = ['学生id','学院编号','成绩排名']
# 导入学习成绩数据
score_train = pd.read_csv("../input/score_train.txt",header=None,encoding='utf-8',names = names_score,index_col = False)
score_test = pd.read_csv("../input/score_test.txt",header=None,encoding='utf-8',names = names_score,index_col = False)
score_data = pd.concat([score_train,score_test])

# 删除后面不用的变量，节省内存
del card_train,card_test,score_train,score_test,names_card,names_score

print(card_data.shape)
print(score_data.shape)

len(card_data.学生id.unique())

# 处理消费方式变量
card_data['消费方式'] = card_data['消费方式'].astype('category')
card_data['消费方式'].describe()

# 看看有多少种消费方式
print(card_data['消费方式'].unique())

数据总共记录了11类消费数据（不含Nan），23998315次消费，其中食堂消费次数最多

# 缺失值处理
# 查看Ｎａｎ
card_data.isnull().sum()

# 由于消费方式中食堂消费最多，使用食堂代替缺失的消费类别
card_data['消费方式'].fillna('食堂',inplace=True)

#每个学生的总消费
card_sum_by_ID = card_data.groupby('学生id')['消费金额'].sum()

card_sum_by_ID.head(20)

## 每个学生各类别的消费
card_sum_by_ID_type = card_data.groupby(['学生id','消费方式'])['消费金额'].sum().unstack('消费方式')

card_sum_by_ID_type.head()

# 缺失值
card_sum_by_ID_type.fillna(0,inplace=True)

card_sum_by_ID_type.head()

## 结合数据
card = pd.concat([card_sum_by_ID,card_sum_by_ID_type],axis =1)

card.head()

del card_sum_by_ID,card_sum_by_ID_type,card_data


## 处理学生学习成绩数据

# 将成绩排名标准化
score_data.成绩排名= score_data.groupby('学院编号').成绩排名.transform(lambda x: (x - x.mean()) / x.std())

score_data.set_index('学生id')

# 保存数据
score_data.to_csv('../input/cleaned_score.csv')
card.to_csv('../input/cleaned_card.csv')

# http://pandas.pydata.org/pandas-docs/stable/io.html#io-read-csv-table
